+++ tests/dummy/app/controllers/index.ts
@@ -40,7 +40,7 @@ import {
 import { placeholder } from '@lblod/ember-rdfa-editor/plugins/placeholder';
 import { inline_rdfa } from '@lblod/ember-rdfa-editor/marks';
 import SayController from '@lblod/ember-rdfa-editor/core/say-controller';
-import { link, linkView } from '@lblod/ember-rdfa-editor/nodes/link';
+import { link, linkView } from '@lblod/ember-rdfa-editor/plugins/link';
 import { inject as service } from '@ember/service';
 import IntlService from 'ember-intl/services/intl';